chore: move `delayed_type_checks` field from `NodeInterner` to `TypeChecker`
